refactor(config): move the online block into its own package
The online block was about half of comicbox/config: all of online.py,
plus ~280 lines of settings.py and the four confuse templates in
__init__.py. It now lives in comicbox/config/online/ as settings.py (the
dataclass tree, enums and resolve_* helpers), template.py (the confuse
schema) and build.py (assembly from the view plus CLI overrides).

What's left in config/ is the general layer at about a thousand lines
across six files, and settings.py reads as the config taxonomy again
rather than mostly online detail.

online/settings.py must not import comicbox.formats at runtime --
formats imports it, so a reference back would close a cycle, and the
deferred imports in config/computed.py and box/init.py exist because of
it. test_config_online_leaf asserts that against the source, since
importing the module also runs the config package __init__ (which does
legitimately import formats.sources) and would mask the real question.

Also deletes config/settings.py's resolve_credentials, which had no
callers and collided by name with the real one in
formats/base/online/credentials.py.
