+++ tests/unit/test_online_matcher.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import pytest
 
-from comicbox.config.settings import (
+from comicbox.config.online.settings import (
     MatchMode,
     OnlineLookupSettings,
     OnlineSettings,
