+++ tests/unit/test_online_selector.py
@@ -10,7 +10,7 @@ import pytest
 
 from comicbox.box import Comicbox
 from comicbox.box.online_lookup import ComicboxOnlineLookup, OnlineLookupAbortedError
-from comicbox.config.settings import MatchMode, Prompts
+from comicbox.config.online.settings import MatchMode, Prompts
 from comicbox.formats import MetadataFormats
 from comicbox.formats.base.online.profile import Candidate, CandidateSummary
 from comicbox.formats.sources import MetadataSources
